Set default prefetch buffer size as long long type.

+++ src/CacheFileInfo.hh
@@ -28,7 +28,8 @@ public:
 
    int getSizeInBytes() const;
    int getSizeInBits() const;
-   long getBufferSize() const;
+
+   long long getBufferSize() const;
    bool testBit(int i) const;
 
    bool isComplete() const;
@@ -39,7 +40,7 @@ public:
 
 
 private:
-   long    m_bufferSize;
+   long long    m_bufferSize;
    time_t m_accessTime;
    int    m_accessCnt;
    int    m_sizeInBits; // number of file blocks
@@ -95,7 +96,7 @@ inline void CacheFileInfo::setBit(int i)
    m_buff[cn] |= cfiBIT(off);
 }
 
-inline long CacheFileInfo::getBufferSize() const
+inline long long CacheFileInfo::getBufferSize() const
 {
    return m_bufferSize;
 }
